Fix macros for reserved prefixes on lifetimes
Rust RFC 3101 reserved, in Rust 2021, prefixed identifiers such as
`prefix#ident`.  This should have been applied to lifetimes before the
release of the edition, but was not.  The Rust Project is planning to
fix this, and this commit prepares this crate for that.

The key change is as follows.  There are many places where this crate
says, in the context of `quote!(..)`:

```rust
impl<'prov, #(#generic_params),*>
```

But there were two places where this was spelled:

```rust
impl<'prov#(,#generic_params)*>
```

It's these that will break, and so this commit changes these to be
consistent with the first spelling.  An alternate way to fix this
would be to just add a space, such as:

```rust
impl<'prov #(,#generic_params)*>
```

...but the first fix seems more idiomatic.

The other changes here aren't strictly necessary, but they make things
more consistent and rely on fewer peculiarities of the Rust parser.